update profile done testing

+++ entities/requests/update_profile_request.go
@@ -3,5 +3,5 @@ package requests
 type UpdateProfileRequest struct {
 	Name     string `json:"name" form:"name" query:"name"`
 	Password string `json:"password" form:"password" query:"password"`
-	Color float32 `json:"color" form:"color" query:"color"`
+	Color    uint   `json:"color" form:"color" query:"color"`
 }
